Remove DB param configuration & force user to specify connection params
Automatically assigning DB params based on the general connection params
is not very explicit functionality and is confusing.

This commit removes that functionality completely, and forces the user
to specify DB connection params directly in the config.

Any duplication of connection parameter strings can be eliminated in the
usual way by using Symfony parameters or environment variables when
configuring this bundle.